requirements-dev.txt
-r requirements.txt
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
django-debug-toolbar==3.1.1
factory-boy==3.2.0